Restructures Test Files
