+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "start": "babel-node server",
     "start:dev": "nodemon --exec babel-node server",
-    "test": "set NODE_ENV=test&& nyc mocha --timeout 500000 --exit server/specs/**/*.js",
+    "test": "set NODE_ENV=test&& nyc mocha --timeout 500000 --exit server/specs",
     "lint": "eslint --fix server",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
